Add gvir_config_domain_set_clock
The implementation is likely to need to be completed later. We
might want to store pointers from GVirConfigDomain to the associated
GVirConfigDomainClock, from GVirConfigDomainClock to the GVirConfigDomain
that contains it. Since I'm not sure yet if they will be needed, I chose
to keep the implementation simple.

--
v2: use g_return_if_fail to test function args for sanity