Add gvir_config_object_add_child
It's similar to gvir_config_object_replace_child except that if the
current node already has a child with the correct name, it returns
the existing child instead of replacing it.

--
v2: instead of adding an argument to gvir_config_object_new_child,
split the function in 2 separate ones, gvir_config_object_add_child
and gvir_config_object_replace_child.
    use g_return_if_fail to test function args for sanity