Add support for creating watches on streams
The GIO GInputStream/GOutputStream async model for I/O does not
work for working with non-blocking bi-directional streams. To
allow that to be done more effectively, add an API to allow
main loop watches to be registered against streams.

Since the libvirt level virStreamEventAddCallback API only allows
a single callback to be registered to a stream at any time, the
GVirStream object needs to be multiplexing of multiple watches into
a single libvirt level callback.

Watches can be removed in the normal way with g_source_remove

* libvirt-gobject/libvirt-gobject-stream.c,
  libvirt-gobject/libvirt-gobject-stream.h,
  libvirt-gobject/libvirt-gobject.sym: Add gvir_stream_add_watch