Fix GIR generation
Overriding PKG_CONFIG_PATH when generating LibvirtGObject-1.0.gir
is wrong since this means that if glib/gio/... are installed in
a prefix that is not one of the default pkg-config paths, then
they won't be found. PKG_CONFIG_PATH is overridden so that
g-ir-scanner can find LibvirtGLib-1.0 and LibvirtGConfig-1.0 in
the build tree. Since g-ir-scanner provides --include-uninstalled
for that purpose, use that instead.