Ensure debug defaults to off for all glib versions
The previous patch changed all code to directly call g_debug().
This relies on the fact that glib >= 2.31 defaults to not
printing debug messages. The result is that for older glib we
now pollute the stderr with debug messages.

This introduces 3 env vars for controlling libvirt debug,
though on glib >= 2.31 the standard G_LOG_MESSAGES env can
be used instead.

* libvirt-glib/libvirt-glib-main.c: Check for LIBVIRT_GLIB_DEBUG
  env var
* libvirt-gobject/libvirt-gobject-main.c: Check for
  LIBVIRT_GOBJECT_DEBUG env var
* libvirt-gconfig/libvirt-gconfig-main.c: Check for
  LIBVIRT_GCONFIG_DEBUG env var
* libvirt-gconfig/libvirt-gconfig-main.h, libvirt-gconfig/libvirt-gconfig.h,
  libvirt-gconfig/libvirt-gconfig.sym: Add new files
* libvirt-gconfig/Makefile.am, libvirt-glib/Makefile.am,
  libvirt-gobject/Makefile.am: Set a default G_LOG_DOMAIN to
  allow filtering the library messages