gir: use libraries from build tree, not installed ones
If we don't explicitly mention path to our uninstalled libraries
to g-ir-scanner, it will use installed libvirt-* libraries instead
of the ones we just built to run a binary during the generation of
the gir file. However this binary will have been compiled with the
uninstalled ones,
This is an issue if the installed libraries are not ABI compatible
with the uninstalled one, and will cause a compile failure because
the binary will fail to run.
This fix has been suggested by Colin Walters (gobject-introspection
maintainer).