Add GObject boilerplate for GVirConfigStorage*
We'll need several helper classes to implement configuration of
storage pools and volumes. This commit introduces all of these
in one go. It's just the GObject boilerplate, the actual code will
come in other commits. GVirConfigStoragePermissions is namespaced this
way because it will be used by both GVirConfigStoragePoolTarget and
GVirConfigStorageVolTarget.