Add gvir_config_object_foreach_child helper
When iterating over xmlNodePtr children to parse an XML document
describing a libvirt configuration item, we want to ignore blank
nodes that may have been added to make the initial XML document
"more human readable". Since this will be useful in several places,
move this code to a helper function.