Follow glib conventions for SimpleAsyncResult source tag
g_simple_async_result_is_valid() API documentation says that the
source tag field for SimpleAsyncResult objects "by convention, is a
pointer to the _async function corresponding to the _finish function from
which this function is called"
The stream functions were already following this convention, but
other places were using the sync function name as the source tag.
This commit uses the async function name everywhere.