Split gvir_config_object_attach
Most of the time we want gvir_config_object_attach to replace
existing nodes, but sometimes (for <devices> subnodes) we want
it to append the new node and to keep the existing nodes with
the same name. This commit solves this by adding 2 distinct helpers,
_attach_add and _attach_replace.
This should fix some unexpected behaviour of various _set_ functions
which were appending new nodes instead of replacing the existing one.