Don't use g_error in xxx_get_name
g_error will trigger an abort of the running process so it's not
desirable to call it in a library. This commit uses g_warning when
this occurs and returns NULL. libvirt will only return NULL for these
strings if some kind of memory corruption is going on, or if we are
using an invalid pointer.