Don't leak GVirStream
Do not give away a reference to virStreamEventAddCallback, but manage
stream life-time and event instead.

It seems to be a bug in current implementation of libvirt
remoteStreamCallbackFree() that doesn't call the free/notify cb for
some reason. (even if it did, I am not sure it would work correctly,
so I prefer that patch)