Add GVirConfigDomainRedirdev
This is used to add the SPICE USB redirection channel. Even if
the libvirt doc doesn't document it with the other devices deriving
from GVirConfigDomainChardev, I think it makes sense to have this
class derivers from GVirConfigDomainChardev too since it needs a
GVirConfigDomainChardevSource, and it's documented as using a
character device for redirection.