Make sure %with_vala is defined for newer RHELs
We want to build the vala bindings on recent RHELs. Given than
%with_introspection is defined for fedora >= 15 and rhel > 6 which
are exactly the versions for which we want to build the vala bindings,
and given that the vala bindings are to be disabled when introspection
is disabled, we can set %with_vala to the same value as %with_introspection.