object: Add "transfer none" annotation to argv parameter
This makes the parameter to be passed "unowned" in Vala. This was
previously done using a vala metadata file, but it's better to do
it directly through a gtk-doc annotation, as this means the gir
file will know about this, and thus any gir-based binding can
make use of this info.

This also makes libvirt-gobject consistent with what was done
for gconfig and glib in commit 431720.