Rename misnamed 'conn' variable everywhere
Through copy and paste from libvirt-gobject-connection.c, we end
up with an instance variable named 'conn' in most of our classes
even if we are not manipulating a connection object. This commit
renames all 'conn' occurrences in libvirt-gobject and
libvirt-gconfig with a name more consistent with the type of the
variable named 'conn'.