gconfig: Add gvir_config_domain_disk_[gs]et_driver_format
libvirt commit
http://libvirt.org/git/?p=libvirt.git;a=commit;h=e2c41e486018ee74f6a75c1f717622
strongly hints that driver type for a disk domain should be an
enum rather than a raw string. Since we already have setters/getters
for this attribute, we can't use the same name as in the XML. However,
as Daniel Berrangé pointed out,
gvir_config_domain_disk_[gs]et_driver_format is a better name for this,
so we add the API accepting an enum under this name.

I did not set "Rename to: gvir_config_domain_disk_[gs]et_driver_format"
in a gtk-doc comment to force bindings to use this better version
as this would break Boxes for example.