Deprecate gvir_config_domain_disk_[gs]et_driver_*
They are replaced by equivalent methods in GVirConfigDomainDiskDriver.
Initially, we had only one or two attributes to set on the 'driver' child
of the 'disk' node. Nowadays, we more than 5 attributes to set on this
node, mapping it as a separate object is more consistent with the rest of
libvirt-gconfig API.