Fix typo in symbol name in libvirt-gobject.sym
gvir_connection_restore_domain_from_file_finish was not properly exported
because of a typo in the symbol name.
This commit also removes a newer occurrence of this symbol in the file
(without the typo). This can be done without breaking ABI as we haven't
made a release since this duplicate symbol was added in the symbol file.