tests: Move EXTRA_DIST back into ENABLE_TESTS conditional
When deciding which files to put into the make dist tarball, automake is
clever enough to ignore conditional blocks, and to pick all files which are
needed for all values of the conditional.

Having EXTRA_DIST=xxx at the beginning of tests/Makefile.am causes
warnings from automake when ENABLE_TESTS is true as glib-tap.mk will
unconditionally empty it.

Keeping EXTRA_DIST+=xxx outside of the ENABLE_TESTS conditional would
cause a different warning from automake when ENABLE_TESTS is false as
we'd be appending to a variable which was never set.

Moving EXTRA_DIST+= back into the conditional solves these 2 warnings.