Require libvirt-glib >= 1.2.6
With bindings added to API that came in 1.2.6:

81c5969 - gobject: Add wrapper for virNetworkGetDHCPLeases
deed153 - gobject: Add wrapper for virNetworkDHCPLease

we better make our requirement explicit.

The alternative would be to make use of #ifdef to only use newer symbols
from libvirt if installed version is >= 1.2.6 but that would require at
least 12 #ifdef in different places and hence very ugly IMO. Keeping in
mind that 1.2.6 is more than an year old now and it doesn't make much
sense to wanting to upgrade to latest libvirt-glib but not wanting to
upgrade to even an year old libvirt, this is very unlikely to cause any
real issues.