README: formally document intended platform support targets
Give users an indication of what distro platforms the project
intends to be buildable on. This policy will be used to decide
when it is appropriate to increase the minimum required versions
of external dependancies.

As an example, as of July 1st 2015, this policy would aim to pick the
min versions based on what's available across the following sample
distros:

 - RHEL: 7.0, 7.1
 - Fedora: 21, 22
 - Ubuntu LTS: 14.04 (Trusty)
 - Ubuntu: 14.10 (Utopic), 15.04 (Vivid)
 - Suse: 12.0
 - OpenSUSE: 13.1, 13.2
 - Debian: 8.0 (Jessie)

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>