gconfig: DomainDevice now instantiable
We'll need to instatiate DomainDevice baseclass itself for unknown (think
new devices added to libvirt XML) devices in a following patch. This
change makes that possible.

This doesn't break any API or ABI to the best of my knowledge and this
assumption was confirmed by Emmanuele Bassi and Tim-Philipp Muller.