scripts: Fix sha-bang lines
This is particularly useful on operating systems that don't ship
Perl or Python as part of the base system (eg. FreeBSD) while still
working just as well as it did before on Linux.

While at it, make it explicit that our scripts are only going to
work with Python 2, add a couple of missing sha-bang lines and
tweak a couple more that just didn't look very nice.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>