Set desired glib min/max API versions
Setting GLIB_VERSION_MAX_ALLOWED ensures we get warnings if
we use APIs that are newer than our declared minimum glib
version.

Setting GLIB_VERSION_MIN_REQUIRED ensures that we get
warnings if we use APIs deprecated in versions preceeding
this.

If the latter is omitted then we get warnings for all
deprecations in glib, which is undesirable if we want to
keep compat with older versions.

This avoids a build error with latest glib

  ./../libvirt-gconfig/libvirt-gconfig-capabilities-cpu-model.c: In function 'gvir_config_capabilities_cpu_model_init':
  ../../libvirt-gconfig/libvirt-gconfig-capabilities-cpu-model.c:44:13: error: G_ADD_PRIVATE [-Werror]
     44 |     model->priv = GVIR_CONFIG_CAPABILITIES_CPU_MODEL_GET_PRIVATE(model);
        |             ^~~~~~~~~~~~~~~

Caused by G_TYPE_INSTANCE_GET_PRIVATE being deprecated.

Reviewed-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>