glib,docs: Remove "object-tree" chapter
gtk-doc-1.30+ does *not* create tree_index.sgml in case the library does
not provide GObjects, which is the case of LibvirtGLib. In previous
versions, though, an empty file would be generated.

This seems to be a regression which hit *a* *lot* of different projects,
all of them having to take the very same approach.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>