po: delete empty translation files
Historically we stored all translation languages known to Zanata, even
if there were no translated strings present, because every language was
created in Zanata. With the move to Weblate we are only going to store
non-empty translations because languages won't be created in Weblate
until someone wants to actually translate them.

Reviewed-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>