po: change update rules to use weblate instead of zanata
The Zanata project is dead and so Fedora is transitioning over to using
Weblate for managing translations. With Weblate, languages are only
created on the server when the first translation is started, so we must
check to see if any new languages exist during download, so that they
can be added to the local languages list.

Reviewed-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>