gitlab: introduce CI jobs testing git master & distro libvirt
The glib build needs to validate two axis

  - A variety of libvirt versions
  - Native and mingw

We test a variety of libvirt versions by running a build against the
distro provided libvirt packages. All that is then missing is a build
against the latest libvirt git master, which only needs to be run on
a single distro, for which CentOS 8 is picked as a stable long life
base.

For mingw, we only test against libvirt git master, using Fedora
rawhide. It is expected that any issues against older libvirt will
have already been caught by the native builds.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>