meson: don't set global project cflags
The add_project_arguments() method causes cflags to be picked
up by g-ir-scanner and gtk-doc tools. This is not desirable as
the code they work with is fully auto-generated and thus not
under our control. We expect the maintainers of those apps to
have generated correct code and don't care if it triggers some
of the aggressive warning flags we are adding for our own code.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>