ci: mark non-released distros as non-gating
The non-released distros like sid, rawhide & tumbleweed are
periodically unstable. As such we don't want them to act as
gating jobs that block merging code, merely advisory.

When marking rawhide as non-gating, we move one of the mingw
jobs to stable Fedora, so we still have some gating coverage
of mingw. The 32-bit job is picked because 32-bit issues are
the most likely to hit.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>